requirements.txt
anyio==4.11.0
APScheduler==3.6.3
cachetools==4.2.2
certifi==2025.10.5
charset-normalizer==3.4.4
coloredlogs==15.0.1
contourpy==1.3.3
cycler==0.12.1
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.60.1
fsspec==2025.10.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
idna==3.11
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
networkx==3.5
numpy==2.1.3
onnxruntime-gpu==1.23.2
opencv-contrib-python-headless==4.10.0.84
opencv-python==4.12.0.88
packaging==25.0
pillow==12.0.0
polars==1.35.1
polars-runtime-32==1.35.1
protobuf==6.33.0
psutil==7.1.2
pyparsing==3.2.5
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-telegram-bot==22.5
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
torch==2.9.0
torchvision==0.24.0
tornado==6.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
ultralytics==8.3.223
ultralytics-thop==2.0.18
urllib3==2.5.0
wheel==0.45.1
